+++ backend/pcloud-api/src/main/java/com/api/config/moduleutils/ComponentScanConfig.java
@@ -1,4 +1,4 @@
-package com.api.config.scan;
+package com.api.config.moduleutils;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
